Added javascript
The input box will display "blank" unless the user clicks in the box,
then it will be blank.  If the user has something typed in the input box
and it loses focus it will retain the text that is in there.  If the
input box is blank, it'll return "blank" to the box.